+++ CHANGELOG.md
@@ -8,6 +8,9 @@
     - The library used to download the database files can be changed by configuring a module implementing the `RefInspector.Downloader.Adapter` behaviour
     - The library used to read YAML files can be changed by using the `:yaml_file_reader` configuration
 
+- Deprecations
+    - The functions `RefInspector.Downloader.path_local/1`, `RefInspector.Downloader.path_remote/1` and `RefInspector.Downloader.read_remote/1` are now declared internal and will result in a `Logger.info/1` message when called until they will be eventually removed
+
 ## v1.0.0 (2018-11-24)
 
 - Ownership has been transferred to the [`elixir-inspector`](https://github.com/elixir-inspector) organisation